work to have emulated arm cortex a15 in cloonix

+++ cloonix/common/agent_dropbear/cloonix_dropbear/cloonix_cli.c
@@ -890,7 +890,7 @@ size_t cloonix_write(int fd, const void *ibuf, size_t count)
     {
     result = write(fd, buf, count);
     if (result != count)
-      KERR("%d %d", result, count);
+      KERR("%d %d", result, (int) count);
     }
   return result;
 }
